ai-classroom-backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15
django-cors-headers>=4.4
channels>=4.1
django-allauth>=65.0
django-extensions>=3.2
PyJWT>=2.8
pdfplumber>=0.11
groq>=0.18.0
pydantic>=2.8
pydantic-settings>=2.0
Pillow>=10.4
python-dotenv>=1.0
chromadb>=1.0
sentence-transformers>=3.0
pypdfium2>=4.30
rapidocr-onnxruntime>=1.2
bleach>=6.1
python-json-logger>=2.0
drf-spectacular>=0.27
redis>=5.0
celery>=5.3
sentry-sdk>=1.40
